change(database): move session_generator() to database folder
