+++ kagami/database/session.py
@@ -1,0 +1,15 @@
+from collections.abc import AsyncGenerator
+from contextlib import asynccontextmanager
+
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from .engine import DatabaseEngine
+
+
+@asynccontextmanager
+async def session_generator() -> AsyncGenerator[AsyncSession]:
+    session: AsyncSession = DatabaseEngine.session_factory()
+    try:
+        yield session
+    finally:
+        await session.close()
